Migration away from Derby DB: infrastructure for parallel DB writes
This is a precursor step to allow writing to both a new database and old.

Summary of changes:
- extracted an interface from DBUserController and named it 'UserDao'.
- moved the static validation methods from DBUserController to the interface 'UserDao'
- took the contents of DBUserController, and moved those contents to a new class 'DerbyUserController', which also impelments 'UserDao'
- Updated DBUserController to delegates all calls to 'DerbyUserController' (DBUserController is now effectively a facade)
- Added capability for primary/secondary datasource:
  - DBUserController uses a UserDaoPrimarySecondary list to find a primary source to read from. DBUserController will log and ignore any errors thrown by the secondary datasource.
  - Added a no-op secondary datasource

What this commit changes:
- we will now write to two datasources, the primary is the current (soon to be legacy) Derby DB datasource. The secondary is a no-op, and will have functionality added in further iterations
- infrastructure only change, no user facing changes

Next steps:
- convert no-op datasource to do something, eventually writing to a datasource
- add some migration logic, on logic, retrieve user records from old datasource and update the new one, re-encrypting the users password.
- add infra to also read from both datasources and to give us a statistic of how much overlap there is
- deploy this configuration
- let some time pass while we wait for the overlap statistic to converge
- switch primary datasources and then simpllify: remove parallel read/write infrastructure, remove non-primary datasource, rename SqlUserController to 'DBUserController'
