+++ src/main/java/games/strategy/engine/lobby/client/login/LoginPanel.java
@@ -26,6 +26,7 @@ import javax.swing.KeyStroke;
 import javax.swing.SwingUtilities;
 
 import games.strategy.engine.lobby.server.userDB.DBUserController;
+import games.strategy.engine.lobby.server.userDB.UserDao;
 import games.strategy.ui.SwingAction;
 import games.strategy.ui.Util;
 
@@ -125,8 +126,8 @@ public class LoginPanel extends JPanel {
   private void logonPressed() {
     final String userName = m_userName.getText();
     final boolean anonymous = m_anonymous.isSelected();
-    if (DBUserController.validateUserName(userName) != null) {
-      JOptionPane.showMessageDialog(this, DBUserController.validateUserName(userName), "Invalid Username",
+    if (UserDao.validateUserName(userName) != null) {
+      JOptionPane.showMessageDialog(this, UserDao.validateUserName(userName), "Invalid Username",
           JOptionPane.ERROR_MESSAGE);
       return;
     } else if (m_password.getPassword().length == 0 && !anonymous) {
