+++ src/main/java/games/strategy/triplea/ui/menubar/LobbyMenu.java
@@ -30,6 +30,7 @@ import games.strategy.engine.lobby.server.IUserManager;
 import games.strategy.engine.lobby.server.ModeratorController;
 import games.strategy.engine.lobby.server.userDB.DBUser;
 import games.strategy.engine.lobby.server.userDB.DBUserController;
+import games.strategy.engine.lobby.server.userDB.UserDao;
 import games.strategy.net.INode;
 import games.strategy.net.Node;
 import games.strategy.sound.SoundOptions;
@@ -149,7 +150,7 @@ public class LobbyMenu extends JMenuBar {
       if (name1 == null || name1.length() < 1) {
         return;
       }
-      if (DBUserController.validateUserName(name1) != null) {
+      if (UserDao.validateUserName(name1) != null) {
         if (JOptionPane.showConfirmDialog(m_frame,
             "The username you entered is invalid. Do you want to ban it anyhow?", "Invalid Username",
             JOptionPane.YES_NO_CANCEL_OPTION) != JOptionPane.YES_OPTION) {
@@ -221,7 +222,7 @@ public class LobbyMenu extends JMenuBar {
       if (name1 == null || name1.length() < 1) {
         return;
       }
-      if (DBUserController.validateUserName(name1) != null) {
+      if (UserDao.validateUserName(name1) != null) {
         if (JOptionPane.showConfirmDialog(m_frame,
             "The username you entered is invalid. Do you want to ban it anyhow?", "Invalid Username",
             JOptionPane.YES_NO_CANCEL_OPTION) != JOptionPane.YES_OPTION) {
